rm old purchase modal reference
