--- components/billing_page/process_payment.css
@@ -1,19 +1,0 @@
-.ProcessPayment-progress {
-    width: 387px;
-    height: 4px;
-    margin: 0 auto;
-    margin-top: 60px;
-    background: rgb(34, 64, 109, 0.12);
-    border-radius: 4px;
-}
-
-.ProcessPayment-progress-fill {
-    height: 100%;
-    background: #0058cc;
-    border-radius: 4px;
-}
-
-.ProcessPayment-body {
-    overflow-x: hidden;
-    overflow-y: hidden;
-}
