+++ components/pricing_modal/content.tsx
@@ -17,8 +17,7 @@ import {
     getCloudProduct as selectCloudProduct,
     getCloudProducts as selectCloudProducts} from 'mattermost-redux/selectors/entities/cloud';
 
-import PurchaseModal from 'components/purchase_modal';
-import BillingPage from 'components/billing_page';
+import BillingPage from 'components/purchase_modal';
 import {makeAsyncComponent} from 'components/async_load';
 import StarMarkSvg from 'components/widgets/icons/star_mark_icon';
 import CheckMarkSvg from 'components/widgets/icons/check_mark_icon';
@@ -177,19 +176,10 @@ function Content(props: ContentProps) {
         isPostTrial = true;
     }
 
-    // const openPurchaseModal = () => {
-    //     trackEvent('cloud_admin', 'click_open_purchase_modal');
-    //     props.onHide();
-    //     dispatch(openModal({
-    //         modalId: ModalIdentifiers.CLOUD_PURCHASE,
-    //         dialogType: PurchaseModal,
-    //     }));
-    // };
-
     const onBillingPage = () => {
         props.onHide();
         dispatch(openModal({
-            modalId: ModalIdentifiers.BILLING_PAGE,
+            modalId: ModalIdentifiers.CLOUD_PURCHASE,
             dialogType: BillingPage,
         }));
     };
